+++ crates/jobs/src/jobs/mod.rs
@@ -1,0 +1,11 @@
+// SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
+//
+// SPDX-License-Identifier: EUPL-1.2
+
+//! # Jobs that can be run in the OpenTalk job execution system
+
+mod event_cleanup;
+mod self_check;
+
+pub use event_cleanup::EventCleanup;
+pub use self_check::SelfCheck;